Add Catalog and ItemsGroups

+++ ProjectWarehouse.Server/Infrastructure/WarehouseLayoutObject.cs
@@ -2,8 +2,8 @@
 
 public class WarehouseLayoutObject
 {
-    public int X {get; set;}
-    public int Y {get; set;}
-    public int Width {get; set;}
-    public int Height {get; set;}
+    public decimal X {get; set;}
+    public decimal Y {get; set;}
+    public decimal Width {get; set;}
+    public decimal Height {get; set;}
 }
